build(release): add release notes preview script
